fix(PinField): fixed incorrect overwrite behavior edge case when default value is provided
